Modify newSimulatorConfiguration() method in datamanager to use floats instead of longs. The SimulatorConfiguration class in persistence uses floats and so does Lems when extracting already anyway

+++ src/main/java/org/geppetto/core/data/DefaultGeppettoDataManager.java
@@ -443,7 +443,7 @@ public class DefaultGeppettoDataManager implements IGeppettoDataManager
 	}
 
 	@Override
-	public ISimulatorConfiguration newSimulatorConfiguration(String simulator, String conversionService, long timestep, long length)
+	public ISimulatorConfiguration newSimulatorConfiguration(String simulator, String conversionService, float timestep, float length)
 	{
 		return new LocalSimulatorConfiguration(0l, simulator, conversionService, timestep, length, new HashMap<String, String>());
 	}
